bug fixes part 2 (both me and radium did testing)

+++ src/main/java/com/hbm/tileentity/machine/TileEntityPWRController.java
@@ -402,7 +402,7 @@ public class TileEntityPWRController extends TileEntityMachineBase implements IG
 		this.typeLoaded = buf.readInt();
 		this.amountLoaded = buf.readInt();
 		this.rodLevel = buf.readDouble();
-		this.rodTarget = buf.readInt();
+		this.rodTarget = buf.readDouble();
 		this.coreHeatCapacity = buf.readLong();
 		tanks[0].deserialize(buf);
 		tanks[1].deserialize(buf);
